Make the sample code same as our plan in prototype
We have decided to make the model loader API consistent with the WebNN API. This leads to significant changes in the API. We also want to offload the work of "model downloading" to users because it will be very easy for them (e.g. using fetch()) but is very complex for us.